[autotest] Don't forget about host-reimages that have completed.
Once a reimaging job is kicked off, the suite job watches it to see which hosts
it starts running on, so it can lock them to reserve them for use.

b51e845c introduced logic to ignore HostQueueEntries in the reimaging
job that were not 'Running'.  This means we also ignore those that are
'Completed'.  This was a problem, because we'd get the list of hosts
used by the reimaging jobs afresh every time we polled for new hosts
to lock.  The code remembered that it locked them, but not that they
were hosts in the job.  Since we loop waiting for the hosts in the job
to match the set of locked hosts, forgetting like this would make us
wait forever.

Instead of fetching the list of hosts in the reimaging job fresh every time,
the code now fetches the list and intelligently merges it into the set
its been keeping track of all along.

BUG=chromium-os:33920
TEST=unit
TEST=run_suite.py, abort reimaging on a single host.  Suite should complete.

Change-Id: I5cfa355d20b27f8a9ed3be3ebad4a7cc086e2f3b
Reviewed-on: https://gerrit.chromium.org/gerrit/33465
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>