faft: Pick the absolute values of the current values from servo and kernel
The signs of the current values from servo and kernel are not consistent
across different devices. So we pick the absolute values.

We got the current values from Snow:
  when AC, servo: -, kernel +
  when DC, servo: +, kernel -

We got the current values from Link:
  when AC, servo: -, kernel +
  when DC, servo: +, kernel +

This is a temporary fix to not break Snow FAFT. Should investigate into
this issue to know more.

BUG=chrome-os-partner:14157
TEST=Run firmware_ECBattery on Snow and Link. AC and DC passed.

Change-Id: Ic1a07361897e90f9801ff9c0ef6000a1537bda47
Reviewed-on: https://gerrit.chromium.org/gerrit/33502
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>