firmware_TouchpadMTB: LinearityValidator: specify a single slot
In the old way, multiple slots (fingers) could be tested in a
single LinearityValidator. The average linearity is determined
by taking into account all of the slots. It was not clear as to
which finger may not be linear enough. In this CL, a
LinearityValidator could only be assigned a slot so that its
linearity and test details could be displayed in a clearer way.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of a chromebook.
$ DISPLAY=:0 sudo python main.py

In the second gesture "two_finger_tracking", the linearity
test results about two fingers are displayed separatedly.

Change-Id: Iefe517be7897e5caf48301d15a92b3e3609da78a
Reviewed-on: https://gerrit.chromium.org/gerrit/34601
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>