firmware_TouchpadMTB: define the group size for Mean Squared Error
This CL fine tunes the partial group size for Mean Squared Error
in calculating the simple linear regression formula. In the old
way, 10% of most deviating points are taken into the partial group.
However, when the finger moves slowly, 10% of the points may
include many non-deviating points which effectively reduce the
calculated error. In this CL, the partial group is defined to be 1
in which case any single point jump would result in a significant
error.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of a chromebook.
$ DISPLAY=:0 sudo python main.py

For the first two gestures, It should run normally.

Change-Id: Ib9f6da2d43ddcdbcfbb304aab9ac0f57f5b70c30
Reviewed-on: https://gerrit.chromium.org/gerrit/34602
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>