firmware_TouchpadMTB: add validator_unittest
This CL adds the validator_unittest and a number of event data files
to conduct unit tests. In order to run the unit tests on a host,
MockTouchpadDevice as well as the device description files on some
platforms are also added to emulate the touchpad device.

BUG=chromium-os:35024
TEST=Run the unittests.
$ cd .../autotest/files/client/site_tests/firmware_TouchpadMTB
$ python tests/validators_unittest.py
The unit tests should all pass.

Change-Id: I31d7cb151803f8d0450ce09a59f9c63150929190
Reviewed-on: https://gerrit.chromium.org/gerrit/35410
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>