autotest: Let DhcpHandlingRules be sensitive to time
Add some logic to let tests set expectations with regards to the time of
arrival of DHCP packets.  These expectations are optional so that you don't
need to anticipate the exact timing of the test.  For instance, this lets us
expect a DHCP address renewal packet around 5 seconds before the end of
a lease.

BUG=chromium-os:33764
TEST=New autotest: https://gerrit.chromium.org/gerrit/#/c/33633/

Change-Id: If51beb50d0416ff501fb24e329a9d3637cd6ad4a
Reviewed-on: https://gerrit.chromium.org/gerrit/33620
Commit-Ready: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>