[autotest] Suite.run_and_wait() still sanity checks the control file list
When we were having issues with using partially-staged builds on the
dev server, we would see a suite "succeed" running only a subset of
its tests.  To determine if this was a race, we added code to re-list
the control files in a suite after we ran all the tests we discovered
at the start of the run.  This means that we need to ask the dev
server for the whole list of files, fetch each one, and process it
again.  That's time consuming and adds a bunch of unneeded dev server
load.

BUG=chromium-os:34557
TEST=unit
STATUS=Fixed

Change-Id: Ibb81fe5d7bfeced07f0a8b2c410af39126d8bede
Reviewed-on: https://gerrit.chromium.org/gerrit/33735
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>