Cleanup logging_UncleanShutdownServer autotest some
Remove the "logging_UncleanShutdown/control" file because it's unnecessary,
and it was confusing because it had the wrong contents.  We do not run
logging_UncleanShutdown directly; it is instead run by the autotest
logging_UncleanShutdownServer.

Also fix the "NAME" in "logging_UncleanShutdownServer/control".  That didn't
actually affect the ability to run the autotest, but I think maybe it's used
in the autotests dashboard or something.

BUG=chromium-os:34769
TEST=Ran logging_UncleanShutdownServer autotest (although chrome crashes in
the middle)

Change-Id: I4e3f408fb88853beeb0ce5a4cebcc85e6d3f075b
Reviewed-on: https://gerrit.chromium.org/gerrit/34056
Commit-Ready: Michael Krebs <mkrebs@chromium.org>
Tested-by: Michael Krebs <mkrebs@chromium.org>
Reviewed-by: Thieu Le <thieule@chromium.org>