Autotest: Automatically Reboot DUT's if Repair Fails
If Repair Fails on Stumpy or Kiev, try power cycling the devices in
order to recover them. The repair control file has been updated to
check if the device has the stumpy or kiev label and it's hostname
begins with 'chromeos' (to indicate it is in the lab) and if so it
will loop over the repair process 6 times, power cycling in between
each try.

BUG=chromium-os:34493
TEST=Ensure that if repair fails it communicates with the RPM
infrastructure and retries.

Change-Id: Idfc0bb1f73fd495f2a087dcd72b110ba62d41341
Reviewed-on: https://gerrit.chromium.org/gerrit/34067
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>