firmware_TouchpadMTB: show prompt on the stdout
When running the tests from a host using ssh connecting to a
chromebook, it is preferable to show the prompt on the stdout
too. This also helps to conduct the tests when a chromebook machine
cannot display GTK windows properly due to some problems.

BUG=chromium-os:35024
TEST=Run the following command from a host using ssh to a chrombook:
$ DISPLAY=:0 sudo python main.py

The prompt would be shown on the stdout.

Change-Id: I1d3d1a1b6e69a8ed515776507e7a84a293ca9ba2
Reviewed-on: https://gerrit.chromium.org/gerrit/35540
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>