firmware_TouchpadMTB: buffer the test result until flushed
After performing the prompted gesture and pressing 'd' to delete
that gesture and to try again, both the test result of the first
gesture and that of the 2nd gesture will be shown on the final
test report. It is cleaner to only display the test result of the
latest one.

This CL buffers the test result until it is explicitly flushed.
Any newer test result will replace the old one.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of a chromebook.
$ DISPLAY=:0 sudo python main.py

After performing the prompted gesture, press 'd' to delete the
gesture and performing the same gesture again. Only the test result
of the latest performed gesture will be displayed on the final report.

Change-Id: If12bdf7960273f3422d1cbfeb881d57e1cda4a87
Reviewed-on: https://gerrit.chromium.org/gerrit/35541
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>