firmware_TouchpadMTB: print report in html format
It is desirable to see both mtplot images and test results and logs
simultaneously in the touchpad firmware test report. This CL adds
a firmware_log module to keep the information in a more structural
way about performed gestures and the logs generated by validators.
It also adds another module report_html to print the images, logs,
and results in the html format. The test scores are printed in
colors based on the score ranges.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of a chromebook.
$ DISPLAY=:0 sudo python main.py

Find the existence of the html report file:
/var/tmp/touchpad_firmware_test/latest/touchpad_firmware_report.html

Change-Id: I2be3f7f72ebfeed8fa84f92c5fa3e099706ac9ad
Reviewed-on: https://gerrit.chromium.org/gerrit/35543
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>