[autotest] Given successfully reimaged machines, run tests if possible.
Currently, we pick at least one machine per HostSpec to reimage.  If
we can't find a machine for a given HostSpec, we declare the
associated tests N/A and don't run them.  If the machine(s) we picked
to satisfy a given HostSpec fail reimaging, we declare those tests in ERROR
and don't run them.

This ignores the fact that, many times, some HostSpecs are subsets of others.
Knowing this, we can do better.  If a HostSpec naively seems unsatisfiable
or doomed, we can check to see if some other machine we found and reimaged
would allow us to run the tests associated with that HostSpec.

Here's what we do:
Choose machines to reimage as normal.  When deciding which
HostSpecs are satisfiable or doomed, though, take into account the other
HostSpecs.  If a HostSpec 'S' is potentially doomed or unsatisfied, but we see
that a HostSpec 'T', which is a superset of S, is good to go, then
consider S ok as well.

BUG=chromium-os:34635
TEST=unit
TEST=run_suite.py -s dummy against two machines, one of which has the bluetooth label.  Abort the reimaging of the bluetooth machine.  dummy_BT should fail to run.
TEST=run_suite.py -s dummy against two machines, one of which has the bluetooth label.  Abort the reimaging of the machine without bluetooth.  All tests should run on the one remaining device.

Change-Id: I0df44c4cacf032cdd206749dbd56310e93e588c0
Reviewed-on: https://gerrit.chromium.org/gerrit/34322
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>