firmware_TouchpadMTB: Fix a bug of get_x_y_multiple_slots and add unit test
When initializing a dictionary over a list of keys to the same given
value especially an empty list or an empty dictionary, a copy of the
value should be made for every key so that they would not point to the
same list or dictionary.

A unit test on get_x_y_multiple_slots() is added to verify the
correctness.

BUG=chromium-os:35024
TEST=Run the following unittests on a host rather than on a chromebook:
$ cd .../autotest/files/client/site_tests/firmware_TouchpadMTB
$ python tests/mtb_unittest.py
It should pass all unit tests.

Change-Id: Ic9906fa265196981f6298381b92b65cd7078a770
Reviewed-on: https://gerrit.chromium.org/gerrit/34869
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>