faft: Change the return type of send_uart_command_get_ouput to a literal struct
We are going to move the method send_uart_command_get_output to servod. FAFT
will call it via a RPC, which only accepts literal structures. The original
return type re.MatchObject is not a literal structure.

This change is to change the return type of send_uart_command_get_ouput to
a list of tuples, each of which contains the entire matched string and all
the subgroups of the match.

This change also restricts the argument regex_list to be an explicit list.

BUG=chromium-os:34787
TEST=run_remote_tests.sh --remote $DUT suite:faft_ec

Change-Id: Ic247fd65109d7ad4b716611b3cc6728665b1309c
Reviewed-on: https://gerrit.chromium.org/gerrit/34609
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>