firmware_TouchpadMTB: add a symbolic link to the latest log directory
This CL adds a symbolic link pointing to the latest log directory so
that it is easier to locate the latest touchpad firmware test report.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of a chromebook.
$ DISPLAY=:0 sudo python main.py

Quit anytime by pressing 'x' after performing a couple of gestures.
Find the 'latest' symbolic link pointing to the latest log directory
in /var/tmp/touchpad_firmware_test/ on the chromebook.

Change-Id: I9181e0456fe79d519c83ab89a74cad9b783a29c9
Reviewed-on: https://gerrit.chromium.org/gerrit/35542
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>