firmware_TouchpadMTB: display html test report file using pyauto
This CL uses pyauto to display the html test report in Chrome
using pyauto. An http server is invoked to serve the local URL
of the test report file.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of a chromebook.
$ DISPLAY=:0 sudo python main.py

After finishing all tests or pressing 'x' to exit in the middle
of the test, it will display the test report in Chrome browser.

Change-Id: I42dcc74642e2ff5e1231bc79884f0014d4f8be49
Reviewed-on: https://gerrit.chromium.org/gerrit/35544
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>