faft: Use the servo controls to issue EC commands
With a change on hdctools: https://gerrit.chromium.org/gerrit/34608

We move the UART command methods from FAFT class to servod class. It makes
FAFT tests still able to run on Autotest lab, in which the autoserv and servod
run in different machines. The servo board is connected to the servod machine
but not the autoserv machine. So we need to move these methods from FAFT to
servod.

CQ-DEPEND=34608
BUG=chromium-os:34787
TEST=manaul, run FAFT with a hdctools side change.
$ run_remote_tests.sh --remote $DUT suite:faft_ec

Change-Id: I4db0e31b8f74078dfaacdeb47d507d19b4a23921
Reviewed-on: https://gerrit.chromium.org/gerrit/34610
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>