firmware_TouchpadMTB: stop power management temporarily during the tests
Because the firmware test window grabs the input focus, the power
manager may dim the screen after a few minutes since the test begins.
This CL stops the power management temporarily so that the screen
keeps bright during the tests.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of a chromebook.
$ DISPLAY=:0 sudo python main.py

During the tests, the screen should not dim.

Change-Id: Ieea5109f154400040d2b5f50c9d75cfdbf034629
Reviewed-on: https://gerrit.chromium.org/gerrit/35545
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>