firmware_TouchpadMTB: wrap gesture label and expand the label width
On some platforms, the gesture name and prompt are too long such
that they are cut in the middle. In this CL, the corresponding
label widget is expanded and the lines are wrapped to display the
text adequately.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of Lumpy. The text
should be displayed properly without being cut in the middle.
$ DISPLAY=:0 sudo python main.py

Change-Id: I18894f177da30b4583fcba12a719bedf749ed24d
Reviewed-on: https://gerrit.chromium.org/gerrit/35816
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>