Reland "Change cleanup phase to only restart the UI instead of rebooting the machine."
This is the same as the original but with one modification -- In server_job,
I've changed when we wipe _uncollected_logs.

Why do I need to do this? The way that the Autotest server runs a cleanup job
is as a Server Job with an empty control file but with cleanup=True. This means
that cleanup runs outside of the regular results dir lifetime of a server job.

Original review in: Ibf33136b14cc29f0c86d4428e09572daaedb8cbe

BUG=chromium-os:34489
TEST=Ran the cleanup phase directly with autoserv with argument
server/control_segments/cleanup. I also started up a local autotest
server and ran a Job which ran through client cleanup phases.

Change-Id: I07525959ea3cb76b1c0ec413b9a968660e92f832
Reviewed-on: https://gerrit.chromium.org/gerrit/34935
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>