Increase power_LoadTest WiFi connection retries
The vast majority of power_LoadTest runs in the lab currently fail due
to WiFi issues. The corresponding setup code contains a comment hinting
that the WiFi connection manager is known to be flaky, and that the
current amount of retries might not be sufficient. This change raises
said amount from 3 to 8 hoping to improve the test's stability in the
lab.

BUG=chromium-os:35120
TEST=Run power_LoadTest manually to ensure that the test still performs
as expected. Long-term success of this change can only be dertermined
after it has been submitted and deployed in the lab for a few weeks.

Change-Id: I707a1f63f9c099e7ee0225ed64958f1cc64dfd2b
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34944
Reviewed-by: Sameer Nanda <snanda@chromium.org>