firmware_TouchpadMTB: display test result with proper font size
On some platforms, the text about the test result may be too long
such that they are cut in the middle. In this CL, we calculate
a proper font size to display the test result based on the dynamic
frame width allocated to the result frame. The display format is
also adjusted to become more compact.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of Lumpy. The test result
should be displayed properly without being cut in the middle.
$ DISPLAY=:0 sudo python main.py

Change-Id: Ia8cacdc1da66d6ca54214103049daa41c2f7ec92
Reviewed-on: https://gerrit.chromium.org/gerrit/35817
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>