firmware_TouchpadMTB: a more detailed report file name
This CL adds the board name, firmware version, and test date to
the file name of the test report. Since the file name is not fixed
as before, there must be some way to pass the file name from the
firmware_TouchpadMTB class to DummyTest derived from pyauto. It
seems that pyauto creates and maintains its own global scope, the
file name cannot be passed as a global variable. As a work-around,
the file name is passed as an environment variable.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of a chromebook. Observe
that the test report name comprises the board name, firmware version,
and test date.
$ DISPLAY=:0 sudo python main.py

Change-Id: I438d0618e7789eb365a63cc26a2fefe78ca06a57
Reviewed-on: https://gerrit.chromium.org/gerrit/35819
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>