Add a JSON file for desktopui_PyAutoPerfTests that maps pyauto test names to perf keys.
This is to help developers who need to know which pyauto tests are associated
with which perf keys.  Since some of the perf key names are created
dynmically by the tests, it is not always easy to figure this out by doing
a static code search.

Another follow-up CL will add new flags to desktopui_PyAutoPerfTests that
will make use of this JSON file.

BUG=chromium-os:24738
TEST=Verified the new file is proper JSON by loading it with python's
     simplejson module.

Change-Id: Id262685a816ac683d839c64b6b81d9aa3396c3d4
Reviewed-on: https://gerrit.chromium.org/gerrit/35080
Reviewed-by: asharif <asharif@chromium.org>
Commit-Ready: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>