firmware_TouchpadMTB: conditions to stop a gesture
On some platforms, a gesture may be stopped prematurely before all
fingers have left the pad. This condition may happen when all fingers
stay stationary such that there are no reported packets within the
predefined timeout interval. In this CL, the touchpad events are
examined to calculate the number of tracking IDs. A gesture is stopped
only when two conditions are met:

  (1) there are no reported packets for a timeout interval, and
  (2) the number of tracking IDs is 0.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of Lumpy.
$ DISPLAY=:0 sudo python main.py

Place a stationary finger on the pad for 20 seconds. Check that the
gesture is not stopped while the finger is still on the pad.

Change-Id: Ib9181f1723c004963ef10d154a5a10b9bbecd789
Reviewed-on: https://gerrit.chromium.org/gerrit/35820
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>