firmware_TouchpadMTB: fixed the prompt wording in the config file
We measure the linearity of one-finger tracking and two-finger
tracking lines in the first two gestures. This CL fixed the wording
by explicitly asking the user to use a ruler.

BUG=chromium-os:35024
TEST=Run the following command in the crosh of Lumpy.
$ DISPLAY=:0 OPTIONS=-s sudo python main.py
It would guide the user to perform the first 2 gestures using a ruler.

Change-Id: Ic98a5414a76c0bf8733853d5b80469324d1b682c
Reviewed-on: https://gerrit.chromium.org/gerrit/36326
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>