Autotest: Have the scheduler fix active and complete queue entries.
Currently as part of the scheduler's 24 hour cleanup, it looks for
entries that are marked as both active=True and complete=True and
if it finds any such entries, it sends out an email that will repeat
daily till the database is fixed.

In order to address this and have the code self correct the problem,
the 24 hour cleanup will now mark the entries as not active as they
are completed.

BUG=chromium-os:35220
TEST=Tested on my local scheduler where I manually set active=Complete=1
     and ensured the error occured and when the fix was put in, it self
     corrected.

Change-Id: I954fb4f93753ae843a0cd3169e96d03d3bc63ba3
Reviewed-on: https://gerrit.chromium.org/gerrit/35292
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>