[autotest] Assume that a 'trivial' HostSpec is satisfied by any appropriate DUT.
Introduce the notion of a 'trivial' HostSpec that is defined only by board and
pool.  When searching for enough hosts to satisfy a list of HostSpecs, if we wind
up needing all |num| hosts we're allowed to use to satisfy non-trivial HostSpecs,
consider that fine.

For example, if a suite has a test that requires bluetooth and one that does not,
and we are only allowed to run on one host, that suite is satisfied by a single
host that has bluetooth.  Before this change we would fail, claiming that we
required at least two hosts: one for the bluetooth test, and one for the other
tests.

BUG=chromium-os:35202
TEST=unit
TEST=run the dummy suite with num=2; all tests should run as before
TEST=run the dummy suite with num=1; all tests should run on a bluetooth machine
TEST=run the dummy suite with num=1, and your only bluetooth DUT locked.  non-BT tests should run.
TEST=run the dummy suite with num=1, and your only NON-bluetooth DUT locked.  all tests should run on a bluetooth machine.

Change-Id: Ic5d9553e05c0b6ec5842b426dbfab7130272313a
Reviewed-on: https://gerrit.chromium.org/gerrit/35431
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>