Clarify failure emails on 'experimental' tests.
This was complicated to reproduce. We run some tests as 'experimental'.
Generally, this implies the test is a bit flaky and we should not act
on it's results as aggressively (e.g. should not close the tree).  This
tag is applied to some tests on some releases.  To reproduce this needed
to find a failed, experimental test within the recent email window.
In the end I forced the behavior a bit.

Then, verified that internal 'experimental' flag is getting set properly
took a little time.  Once this was all checked, the code to improve the
email was surprisingly simple.

BUG=chromium-os:35261
TEST=run_generate.py -j 300 -k -m -v
TEST=sample email: https://www/~truty/dashboard/2012-10-12_Experimental_Failure_Email.png

Change-Id: I80b06b7d7f2b5c70206aedf563fced43a87f5ce6
Reviewed-on: https://gerrit.chromium.org/gerrit/35434
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Mike Truty <truty@chromium.org>
Commit-Ready: Mike Truty <truty@chromium.org>