desktopui_PyAutoPerfTests now passes along pyauto error messages to error.TestFail().
desktopui_PyAutoPerfTests is an autotest wrapper around a set of pyauto-based
performance tests.  When the autotest fails, that typically means at least
one of the individual pyauto tests failed.  However, the pyauto error messages
do not show up in the autotest client log or the "status" file that is created after
an autotest runs.  Instead, a user must manually view the stdout DEBUG file
associated with the autotest run to see which pyauto tests failed, and why.

This change makes desktopui_PyAutoPerfTests output the pyauto test failure
messages as part of the error.TestFail() autotest error message.  This causes
the pyauto errors -- if present -- to appear in the autotest client log and in the
"status" files.  However, the pyauto errors do not yet appear in the
output of run_remote_tests.sh, due to an issue where run_remote_tests.sh
only displays the results of the first line of text from the error.TestFail()
message.  A subsequent CL will address that problem so that we can see the
pyauto error messages as part of the output of run_remote_tests.sh too.

BUG=chromium-os:33210
TEST=Verified that with these changes, the results of desktopui_PyAutoPerfTests
     show the individual pyauto error messages in the autotest client DEBUG
     file and in the "status" file.

Change-Id: I55f15a6d74a72d63dd0a7ade22a52293e38c3c9f
Reviewed-on: https://gerrit.chromium.org/gerrit/35608
Reviewed-by: asharif <asharif@chromium.org>
Commit-Ready: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>