Improve error logging in Autotest package system
Autotest uses wget to fetch test packages from the server, but fails to
log its output or error code on failure. This patch improves that
behavior in hopes of shedding more light on some recent power_LoadTest
lab failures.

BUG=chromium-os:35344
TEST=Edit the /build/<board>/usr/local/autotest/global_config.ini file
on the server and add the line 'fetch_location: http://doesnotexist' to
the [PACKAGES] section. Run power_LoadTest and check for the wget error
output and exit code in the stored client.0.DEBUG log.

Change-Id: I02d34786fc0a2d032f535268ddc7ada2256a4d48
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35705
Reviewed-by: Chris Sosa <sosa@chromium.org>