faft: Start devserver to serve the image and ask Servod to dump it to USB
Originally we assume Servod runs in the same machine as Autoserv such that
we can directly install image by giving a local path. But it is not true
in Autotest lab. Servod runs in a different host and doesn't understand
the local path.

We should start a devserver to serve the image, which is similar to the
lab, and pass the URL to Servod. Then Servod directly dump it to its USB.

The image argument also supports an URL now. So change the comment.

If Servod and Autoserv run in the same machine. It increases ~1 minute in
addition after this change (~3min vs ~4min). It takes time to setup devserver
and transfer the data on network loop.

BUG=chrome-os-partner:14884,chromium-os:34395
TEST=manual

For a local file:
$ ./run_remote_tests.sh --board daisy --remote dut -a "image=/home/waihong/trunk
/src/build/images/daisy/R22-2723.88.0/chromiumos_test_image.bin" FAFTSetup
See:
19:30:00 INFO | Server proxy: http://localhost:9990
19:30:11 INFO | Processing build: 2723.88.0 dev-channel
19:30:11 INFO | Starting devserver to serve the image...
19:30:11 INFO | Ask Servo to install the image from http://10.0.0.1:8090/static/
archive/chromiumos_test_image.bin
...
19:30:17 INFO | Searching for usb device and copying image to it.
...

For an URL:
$ start_devserver
$ wget -O - -t 1 -q --timeout 0 'http://localhost:8080/stage_images?archive_url=
gs://chromeos-image-archive/daisy-release/R23-2913.84.0&image_types=test'
$ ./run_remote_tests.sh --board daisy --remote dut -a "image=http://localhost:80
80/static/images/daisy-release/R23-2913.84.0/chromiumos_test_image.bin" FAFTSetup

Change-Id: Iadf30d892e665f3a2898fb31860992b95caa402f
Reviewed-on: https://gerrit.chromium.org/gerrit/35689
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>