Autotest: Add label detection to site_host and add it to the CLI
Currently we manually maintain the list of labels for a given DUT,
in order to make this process easier we added methods to site_hosts
to automatically determine certain attributes/labels for a given
host.

The autotest CLI was also updated such that it can now utilize the
auto label detection methods. If no platform is provided for a given
list of hostnames, it will look up the correct platform for each.
And if it can retrieve the host's labels it will assign a Set of the
detectable labels merged with the command line inputted labels.

BUG=chromium-os:18976
TEST=Ran cli/atest create host locally.

Change-Id: I50480fcddbce60b8d5693f1268ca36d1aea12e23
Reviewed-on: https://gerrit.chromium.org/gerrit/35889
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>