Autotest: Hosts or Host Queue Entry in incorrect states fix.
Currently Host Queue Entries will only start if they are in a correct
state and their hosts are also in a correct state. Discrepancies can
occur if the scheduler dies poorly or if changes to the database don't
'stick'. This can cause a scheduler crashing loop and will impose
downtime in the lab.

In order to prevent this, I have updated the codepath to log and email
an error message when this occurs and just abort the job attached to the
Host/Host Queue Entry. This will prevent the scheduler from going down,
and will allow us to keep processing other jobs while taking a hit of
losing one job's results.

BUG=chromium-os:35366
TEST=On my local machine by recreating the db inconsistency and witnessing
     monitordb fail to launch, applying the fix and seeing the job complete.

Change-Id: I24c5eaca5a25fa869ee69570044929b9c98332c6
Reviewed-on: https://gerrit.chromium.org/gerrit/35743
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>