autotest: Add module: capture packets on a specific frequency/ht setup.
The Chaos Lab is being setup to test why a DUT may not be connecting to
a specific AP.  This module captures packets in the background on the
frequency/HT that the DUT is trying to use.

For this to work, in this form, at least one machine needs to be added
to the Autotest machine pool with 'packet_capture' label assigned.

BUG=chromium-os:35400
TEST=Manual.

Change-Id: Id196dbcf0b43db750de16874af206b3cba2fec6f
Reviewed-on: https://gerrit.chromium.org/gerrit/37128
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>