lab_test: Support image_file in RO directory for VerifyImageAndGetId
A bit refactoring on FAFT will re-use VerifyImageAndGetId to an image_file
in RO directory, like /dev/sdc.

Originally, it created the directories for mount in the same directory of
image_file. It failed in the RO case. So this CL is to make it work by
creating a new temporary directory.

BUG=chromium-os:34395
TEST=manual, run the following test which calls VerifyImageAndGetId.
$ run_remote_tests.sh --board daisy --remote dut RecoveryButton/control$

Change-Id: I3de68caa15f788189e590f351f2ed50a8e6872a4
Reviewed-on: https://gerrit.chromium.org/gerrit/36050
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>