firmware_TouchpadMTB: create key press events in robot mode
When operating in robot mode, the system would command the robot
to perform a gesture at a time. After a gesture is performed, the
system would wait a key press event to determine if it should go to
next gesture, re-perform the gesture again, or quit the system.
In the robot mode, we assume that the robot would perform the gesture
correctly and hence the SAVE key press event is sent to notify the
system to save the current gesture and move forward to next gesture.

BUG=chromium-os:35751
TEST=None. Will be tested with patches later when the gesture list
to test for robot mode is defined.

Change-Id: Ib9a24ab06f8791e62055ae23f6d3fdf4858c4b63
Reviewed-on: https://gerrit.chromium.org/gerrit/36657
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>