firmware_TouchpadMTB: define gesture lists for both manual and robot modes
In this CL, the gesture lists for both manual mode and robot mode are
defined. The original gesture_list in the config file is re-defined
as a gesture dictionary. A proper gesture list is selected based on
the command mode option.

BUG=chromium-os:35751
TEST=Run the following command in the crosh of a chromebook set up
with the touch robot.
$ DISPLAY=:0 OPTIONS="-m ROBOT" python main.py
It would command the robot to perform a series of gestures, and
display the test result on the Chrome browser when all tests are
finished.

Change-Id: Ib618ba790a23ad23fb084cf6b1e1b8a406473f21
Reviewed-on: https://gerrit.chromium.org/gerrit/36658
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>