Autotest: If power_on fails at the end of cleanup, don't fail.
Now that the RPM infrastructure is up, it will be called as part of
the power tests and at the end of cleanup to set the DUT's outlet to
on.

If for some reason the RPM has not been properly configured, can't be
reached or an issue with the RPM infrastructure, this will cause all
our cleanup jobs to start failing.

Therefore for the end of cleanup we will catch and log the error and
we can rely on the RPM Infrastructure emails to notify us of issues.

BUG=None
TEST=Launched cleanup locally.

Change-Id: Id939c2a0e205e2f390a4ce6e2d95ca87f8eeb254
Reviewed-on: https://gerrit.chromium.org/gerrit/36295
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>