Refactor power testing library
This is a refactoring patch trying to deduplicate code between the
several similar state-residency-over-time statistics classes in
cros/power_status.py. It moves as much functionality as possible into a
common superclass. Except for minor formatting changes to
power_LoadTest's output, it does not introduce any functional changes.

BUG=none
TEST=Run power tests on all platforms and compare results to prior
values to ensure they are still as stable and accurate as before.

Change-Id: I0b55b6967e5af1acb0ab13812e4706204ed6999f
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36609
Reviewed-by: Sameer Nanda <snanda@chromium.org>